requirements.txt
# ============================================================
# Energy Management System - Python Dependencies
# ============================================================
# Install:  python -m pip install -r requirements.txt
# ============================================================

# --- Kafka Producer ---
kafka-python==2.0.2

# --- InfluxDB 1.8 Client ---
influxdb==5.3.2

# --- PySpark (Structured Streaming) ---
pyspark==3.5.0

# --- AI / Machine Learning (Prophet changed to Holt-Winters) ---
statsmodels>=0.13.0
pandas>=2.0.0

# --- Keyboard listener for anomaly trigger ---
# OPTIONAL: Only needed on Linux/Mac if select+tty approach fails.
# On Windows, msvcrt (built-in) is used automatically — no install needed.
# On Linux/Mac, this lib requires: sudo pip install keyboard
# keyboard==0.13.5
